feat: module transactions
Moved the transactions to separate module.

Changes to be committed:
	new file:   src/lib/transactions.ts
    - feat: interface TransactionOptions
    - feat: createTransactionId
    - feat: createNewSavepointName
    - feat: startTransaction
    - feat: saveTransaction
    - feat: releaseSavepoint
    - feat: rollbackTransaction
    - feat: commitTransaction
	modified:   src/lib/auth.ts
    - chore: use transcations module.
        - remove: startTransaction
        - remove: saveTransaction
        - remove: releaseSavepoint
        - remove: rollbackTransaction
        - remove: commitTransaction
        - remove: createTransactionId
        - revove: createNewSavepointName
        - remove: interface TransactionOptions
	modified:   src/lib/db.ts
    - chore: use transcations module.

	modified:   src/lib/users.ts
    - chore: use transcations module.
