+++ src/lib/auth.ts
@@ -4,7 +4,7 @@ import { Client, DatabaseError, PoolClient } from 'pg';
 import { randomBytes, pbkdf2, timingSafeEqual, randomUUID, pbkdf2Sync } from 'node:crypto';
 import { Console, ConsoleConstructorOptions } from 'node:console';
 import { safeRelease } from './api_db';
-import { commitTransaction, releaseSavepoint, rollbackTransaction, startTransaction, TransactionOptions } from './db';
+import { commitTransaction, releaseSavepoint, rollbackTransaction, startTransaction, TransactionOptions } from './transactions';
 
 class DebugConsole extends Console {
